drm: Shrink mode->private_flags
gma500 needs 4 bits (to store a pixel multiplier) in the
mode->private_flags, i915 currently has three bits defined.
No one else uses this. Reduce the size to u8.

+++ include/drm/drm_modes.h
@@ -378,7 +378,7 @@ struct drm_display_mode {
 	 * by atomic drivers since they can store any additional data by
 	 * subclassing state structures.
 	 */
-	int private_flags;
+	u8 private_flags;
 
 	/**
 	 * @picture_aspect_ratio:
